Crate imgui_sys

Structs

CustomRect
Font atlas custom rectangle

GlyphRangesBuilder
Font atlas glyph range builder

ImDrawChannel
Temporary storage for outputting drawing commands out of order

ImDrawCmd
A single draw command within a parent ImDrawList (generally maps to 1 GPU draw call)

ImDrawCornerFlags
Flags for indictating which corner of a rectangle should be rounded

ImDrawData
All draw command lists required to render the frame

ImDrawList
A single draw command list (generally one per window)

ImDrawListFlags
Draw list flags

ImDrawListSharedData
Data shared among multiple draw lists

ImDrawVert
A single vertex

ImFont
Runtime data for a single font within a parent ImFontAtlas

ImFontAtlas
Runtime data for multiple fonts, bake multiple fonts into a single texture, TTF/OTF font loader

ImFontAtlasFlags
Flags for font atlases

ImFontConfig
Configuration data when adding a font or merging fonts

ImFontGlyph
Font glyph

ImGuiBackendFlags
Back-end capability flags

ImGuiColorEditFlags
Color edit flags

ImGuiComboFlags
Flags for combo boxes

ImGuiCond
Condition flags

ImGuiConfigFlags
Configuration flags

ImGuiDragDropFlags
Flags for igBeginDragDropSource(), igAcceptDragDropPayload()

ImGuiFocusedFlags
Flags for window focus checks

ImGuiHoveredFlags
Flags for hover checks

ImGuiIO
Main configuration and I/O between your application and ImGui

ImGuiInputTextCallbackData
Shared state of input text callback

ImGuiInputTextFlags
Flags for text inputs

ImGuiListClipper
Helper to manually clip large list of items

ImGuiPayload
Data payload for Drag and Drop operations

ImGuiSelectableFlags
Flags for selectables

ImGuiSizeCallbackData
Callback data for size constraint callback

ImGuiStorage
Key->value storage

ImGuiStyle
Runtime data for styling/colors

ImGuiTextBuffer
Text buffer for logging/accumulating text

ImGuiTextFilter
Parse and apply text filters

ImGuiTreeNodeFlags
Flags for trees and collapsing headers

ImGuiWindowFlags
Window flags

ImVec2
A tuple of 2 floating-point values

ImVec4
A tuple of 4 floating-point values

ImVector
Lightweight vector struct

Pair
ImGuiStorage key->value pair

TextRange
ImGuiTextFilter text range

Enums

ImGuiCol
A color identifier for styling

ImGuiContext
ImGui context (opaque)

ImGuiDataType
A primary data type

ImGuiDir
A cardinal direction

ImGuiKey
A key identifier (ImGui-side enum)

ImGuiMouseCursor
A mouse cursor identifier

ImGuiNavInput
An input identifier for navigation

ImGuiStyleVar
A variable identifier for styling

Constants

ImGuiCol_COUNT	[
Deprecated
]

ImGuiKey_COUNT	[
Deprecated
]

ImGuiMouseCursor_COUNT	[
Deprecated
]

ImGuiStyleVar_COUNT	[
Deprecated
]

Functions

CustomRect_IsPacked⚠

GlyphRangesBuilder_AddChar⚠

GlyphRangesBuilder_AddRanges⚠

GlyphRangesBuilder_AddText⚠

GlyphRangesBuilder_BuildRanges⚠

GlyphRangesBuilder_GetBit⚠

GlyphRangesBuilder_SetBit⚠

ImDrawData_Clear⚠

ImDrawData_DeIndexAllBuffers⚠

ImDrawData_ScaleClipRects⚠

ImDrawList_AddBezierCurve⚠

ImDrawList_AddCallback⚠

ImDrawList_AddCircle⚠

ImDrawList_AddCircleFilled⚠

ImDrawList_AddConvexPolyFilled⚠

ImDrawList_AddDrawCmd⚠

ImDrawList_AddImage⚠

ImDrawList_AddImageQuad⚠

ImDrawList_AddImageRounded⚠

ImDrawList_AddLine⚠

ImDrawList_AddPolyLine⚠

ImDrawList_AddQuad⚠

ImDrawList_AddQuadFilled⚠

ImDrawList_AddRect⚠

ImDrawList_AddRectFilled⚠

ImDrawList_AddRectFilledMultiColor⚠

ImDrawList_AddText⚠

ImDrawList_AddTextFontPtr⚠

ImDrawList_AddTriangle⚠

ImDrawList_AddTriangleFilled⚠

ImDrawList_ChannelsMerge⚠

ImDrawList_ChannelsSetCurrent⚠

ImDrawList_ChannelsSplit⚠

ImDrawList_Clear⚠

ImDrawList_ClearFreeMemory⚠

ImDrawList_CloneOutput⚠

ImDrawList_GetClipRectMax⚠	[
Deprecated
]

ImDrawList_GetClipRectMax_nonUDT2⚠

ImDrawList_GetClipRectMin⚠	[
Deprecated
]

ImDrawList_GetClipRectMin_nonUDT2⚠

ImDrawList_PathArcTo⚠

ImDrawList_PathArcToFast⚠

ImDrawList_PathBezierCurveTo⚠

ImDrawList_PathClear⚠

ImDrawList_PathFillConvex⚠

ImDrawList_PathLineTo⚠

ImDrawList_PathLineToMergeDuplicate⚠

ImDrawList_PathRect⚠

ImDrawList_PathStroke⚠

ImDrawList_PopClipRect⚠

ImDrawList_PopTextureID⚠

ImDrawList_PrimQuadUV⚠

ImDrawList_PrimRect⚠

ImDrawList_PrimRectUV⚠

ImDrawList_PrimReserve⚠

ImDrawList_PrimVtx⚠

ImDrawList_PrimWriteIdx⚠

ImDrawList_PrimWriteVtx⚠

ImDrawList_PushClipRect⚠

ImDrawList_PushClipRectFullScreen⚠

ImDrawList_PushTextureID⚠

ImDrawList_UpdateClipRect⚠

ImDrawList_UpdateTextureID⚠

ImFontAtlas_AddCustomRectFontGlyph⚠

ImFontAtlas_AddCustomRectRegular⚠

ImFontAtlas_AddFont⚠

ImFontAtlas_AddFontDefault⚠

ImFontAtlas_AddFontFromFileTTF⚠

ImFontAtlas_AddFontFromMemoryCompressedBase85TTF⚠

ImFontAtlas_AddFontFromMemoryCompressedTTF⚠

ImFontAtlas_AddFontFromMemoryTTF⚠

ImFontAtlas_Build⚠

ImFontAtlas_CalcCustomRectUV⚠

ImFontAtlas_Clear⚠

ImFontAtlas_ClearFonts⚠

ImFontAtlas_ClearInputData⚠

ImFontAtlas_ClearTexData⚠

ImFontAtlas_GetCustomRectByIndex⚠

ImFontAtlas_GetGlyphRangesChineseFull⚠

ImFontAtlas_GetGlyphRangesChineseSimplifiedCommon⚠

ImFontAtlas_GetGlyphRangesCyrillic⚠

ImFontAtlas_GetGlyphRangesDefault⚠

ImFontAtlas_GetGlyphRangesJapanese⚠

ImFontAtlas_GetGlyphRangesKorean⚠

ImFontAtlas_GetGlyphRangesThai⚠

ImFontAtlas_GetMouseCursorTexData⚠

ImFontAtlas_GetTexDataAsAlpha8⚠

ImFontAtlas_GetTexDataAsRGBA32⚠

ImFontAtlas_IsBuilt⚠

ImFontAtlas_SetTexID⚠

ImFont_AddGlyph⚠

ImFont_AddRemapChar⚠

ImFont_BuildLookupTable⚠

ImFont_CalcTextSizeA⚠	[
Deprecated
]

ImFont_CalcTextSizeA_nonUDT2⚠

ImFont_CalcWordWrapPositionA⚠

ImFont_ClearOutputData⚠

ImFont_FindGlyph⚠

ImFont_FindGlyphNoFallback⚠

ImFont_GetCharAdvance⚠

ImFont_GetDebugName⚠

ImFont_GrowIndex⚠

ImFont_IsLoaded⚠

ImFont_RenderChar⚠

ImFont_RenderText⚠

ImFont_SetFallbackChar⚠

ImGuiIO_AddInputCharacter⚠

ImGuiIO_AddInputCharactersUTF8⚠

ImGuiIO_ClearInputCharacters⚠

ImGuiInputTextCallbackData_DeleteChars⚠

ImGuiInputTextCallbackData_HasSelection⚠

ImGuiInputTextCallbackData_InsertChars⚠

ImGuiListClipper_Begin⚠

ImGuiListClipper_End⚠

ImGuiListClipper_Step⚠

ImGuiPayload_Clear⚠

ImGuiPayload_IsDataType⚠

ImGuiPayload_IsDelivery⚠

ImGuiPayload_IsPreview⚠

ImGuiStorage_BuildSortByKey⚠

ImGuiStorage_Clear⚠

ImGuiStorage_GetBool⚠

ImGuiStorage_GetBoolRef⚠

ImGuiStorage_GetFloat⚠

ImGuiStorage_GetFloatRef⚠

ImGuiStorage_GetInt⚠

ImGuiStorage_GetIntRef⚠

ImGuiStorage_GetVoidPtr⚠

ImGuiStorage_GetVoidPtrRef⚠

ImGuiStorage_SetAllInt⚠

ImGuiStorage_SetBool⚠

ImGuiStorage_SetFloat⚠

ImGuiStorage_SetInt⚠

ImGuiStorage_SetVoidPtr⚠

ImGuiStyle_ScaleAllSizes⚠

ImGuiTextBuffer_appendf⚠

ImGuiTextBuffer_begin⚠

ImGuiTextBuffer_c_str⚠

ImGuiTextBuffer_clear⚠

ImGuiTextBuffer_empty⚠

ImGuiTextBuffer_end⚠

ImGuiTextBuffer_reserve⚠

ImGuiTextBuffer_size⚠

ImGuiTextFilter_Build⚠

ImGuiTextFilter_Clear⚠

ImGuiTextFilter_Draw⚠

ImGuiTextFilter_IsActive⚠

ImGuiTextFilter_PassFilter⚠

TextRange_begin⚠

TextRange_empty⚠

TextRange_end⚠

TextRange_split⚠

igAcceptDragDropPayload⚠

igAlignTextToFramePadding⚠

igArrowButton⚠

igBegin⚠

igBeginChild⚠

igBeginChildFrame⚠

igBeginChildID⚠

igBeginCombo⚠

igBeginDragDropSource⚠
Call when current ID is active.

igBeginDragDropTarget⚠

igBeginGroup⚠

igBeginMainMenuBar⚠

igBeginMenu⚠

igBeginMenuBar⚠

igBeginPopup⚠

igBeginPopupContextItem⚠

igBeginPopupContextVoid⚠

igBeginPopupContextWindow⚠

igBeginPopupModal⚠

igBeginTooltip⚠

igBullet⚠

igBulletText⚠

igButton⚠

igCalcItemWidth⚠

igCalcListClipping⚠

igCalcTextSize⚠	[
Deprecated
]

igCalcTextSize_nonUDT2⚠

igCaptureKeyboardFromApp⚠

igCaptureMouseFromApp⚠

igCheckbox⚠

igCheckboxFlags⚠

igCloseCurrentPopup⚠

igCollapsingHeader⚠

igCollapsingHeaderBoolPtr⚠

igColorButton⚠

igColorConvertFloat4ToU32⚠

igColorConvertHSVtoRGB⚠

igColorConvertRGBtoHSV⚠

igColorConvertU32ToFloat4⚠	[
Deprecated
]

igColorConvertU32ToFloat4_nonUDT2⚠

igColorEdit3⚠

igColorEdit4⚠

igColorPicker3⚠

igColorPicker4⚠

igColumns⚠

igCombo⚠

igComboFnPtr⚠

igComboStr⚠

igCreateContext⚠

igDebugCheckVersionAndDataLayout⚠

igDestroyContext⚠

igDragFloat⚠

igDragFloat2⚠

igDragFloat3⚠

igDragFloat4⚠

igDragFloatRange2⚠

igDragInt⚠

igDragInt2⚠

igDragInt3⚠

igDragInt4⚠

igDragIntRange2⚠

igDragScalar⚠

igDragScalarN⚠

igDummy⚠

igEnd⚠

igEndChild⚠

igEndChildFrame⚠

igEndCombo⚠

igEndDragDropSource⚠

igEndDragDropTarget⚠

igEndFrame⚠

igEndGroup⚠

igEndMainMenuBar⚠

igEndMenu⚠

igEndMenuBar⚠

igEndPopup⚠

igEndTooltip⚠

igGetClipboardText⚠

igGetColorU32⚠

igGetColorU32Vec⚠

igGetColorU32U32⚠

igGetColumnIndex⚠

igGetColumnOffset⚠

igGetColumnWidth⚠

igGetColumnsCount⚠

igGetContentRegionAvail⚠	[
Deprecated
]

igGetContentRegionAvailWidth⚠

igGetContentRegionAvail_nonUDT2⚠

igGetContentRegionMax⚠	[
Deprecated
]

igGetContentRegionMax_nonUDT2⚠

igGetCurrentContext⚠

igGetCursorPos⚠	[
Deprecated
]

igGetCursorPosX⚠

igGetCursorPosY⚠

igGetCursorPos_nonUDT2⚠

igGetCursorScreenPos⚠	[
Deprecated
]

igGetCursorScreenPos_nonUDT2⚠

igGetCursorStartPos⚠	[
Deprecated
]

igGetCursorStartPos_nonUDT2⚠

igGetDragDropPayload⚠

igGetDrawData⚠

igGetDrawListSharedData⚠

igGetFont⚠

igGetFontSize⚠

igGetFontTexUvWhitePixel⚠	[
Deprecated
]

igGetFontTexUvWhitePixel_nonUDT2⚠

igGetFrameCount⚠

igGetFrameHeight⚠

igGetFrameHeightWithSpacing⚠

igGetIDPtr⚠

igGetIDStr⚠

igGetIDStrStr⚠

igGetIO⚠

igGetItemRectMax⚠	[
Deprecated
]

igGetItemRectMax_nonUDT2⚠

igGetItemRectMin⚠	[
Deprecated
]

igGetItemRectMin_nonUDT2⚠

igGetItemRectSize⚠	[
Deprecated
]

igGetItemRectSize_nonUDT2⚠

igGetKeyIndex⚠

igGetKeyPressedAmount⚠

igGetMouseCursor⚠

igGetMouseDragDelta⚠	[
Deprecated
]

igGetMouseDragDelta_nonUDT2⚠

igGetMousePos⚠	[
Deprecated
]

igGetMousePosOnOpeningCurrentPopup⚠	[
Deprecated
]

igGetMousePosOnOpeningCurrentPopup_nonUDT2⚠

igGetMousePos_nonUDT2⚠

igGetOverlayDrawList⚠

igGetScrollMaxX⚠

igGetScrollMaxY⚠

igGetScrollX⚠

igGetScrollY⚠

igGetStateStorage⚠

igGetStyle⚠

igGetStyleColorName⚠

igGetStyleColorVec4⚠

igGetTextLineHeight⚠

igGetTextLineHeightWithSpacing⚠

igGetTime⚠

igGetTreeNodeToLabelSpacing⚠

igGetVersion⚠

igGetWindowContentRegionMax⚠	[
Deprecated
]

igGetWindowContentRegionMax_nonUDT2⚠

igGetWindowContentRegionMin⚠	[
Deprecated
]

igGetWindowContentRegionMin_nonUDT2⚠

igGetWindowContentRegionWidth⚠

igGetWindowDrawList⚠

igGetWindowHeight⚠

igGetWindowPos⚠	[
Deprecated
]

igGetWindowPos_nonUDT2⚠

igGetWindowSize⚠	[
Deprecated
]

igGetWindowSize_nonUDT2⚠

igGetWindowWidth⚠

igImage⚠

igImageButton⚠

igIndent⚠

igInputDouble⚠

igInputFloat⚠

igInputFloat2⚠

igInputFloat3⚠

igInputFloat4⚠

igInputInt⚠

igInputInt2⚠

igInputInt3⚠

igInputInt4⚠

igInputScalar⚠

igInputScalarN⚠

igInputText⚠

igInputTextMultiline⚠

igInvisibleButton⚠

igIsAnyItemActive⚠

igIsAnyItemFocused⚠

igIsAnyItemHovered⚠

igIsAnyMouseDown⚠

igIsItemActive⚠

igIsItemClicked⚠

igIsItemDeactivated⚠

igIsItemDeactivatedAfterEdit⚠

igIsItemEdited⚠

igIsItemFocused⚠

igIsItemHovered⚠

igIsItemVisible⚠

igIsKeyDown⚠

igIsKeyPressed⚠

igIsKeyReleased⚠

igIsMouseClicked⚠

igIsMouseDoubleClicked⚠

igIsMouseDown⚠

igIsMouseDragging⚠

igIsMouseHoveringRect⚠

igIsMousePosValid⚠

igIsMouseReleased⚠

igIsPopupOpen⚠

igIsRectVisible⚠

igIsRectVisibleVec2⚠

igIsWindowAppearing⚠

igIsWindowCollapsed⚠

igIsWindowFocused⚠

igIsWindowHovered⚠

igLabelText⚠

igListBoxFnPtr⚠

igListBoxFooter⚠

igListBoxHeaderInt⚠

igListBoxHeaderVec2⚠

igListBoxStr_arr⚠

igLoadIniSettingsFromDisk⚠

igLoadIniSettingsFromMemory⚠

igLogButtons⚠

igLogFinish⚠

igLogText⚠

igLogToClipboard⚠

igLogToFile⚠

igLogToTTY⚠

igMemAlloc⚠

igMemFree⚠

igMenuItemBool⚠

igMenuItemBoolPtr⚠

igNewFrame⚠

igNewLine⚠

igNextColumn⚠

igOpenPopup⚠

igOpenPopupOnItemClick⚠

igPlotHistogramFloatPtr⚠

igPlotHistogramFnPtr⚠

igPlotLines⚠

igPlotLinesFnPtr⚠

igPopAllowKeyboardFocus⚠

igPopButtonRepeat⚠

igPopClipRect⚠

igPopFont⚠

igPopID⚠

igPopItemWidth⚠

igPopStyleColor⚠

igPopStyleVar⚠

igPopTextWrapPos⚠

igProgressBar⚠

igPushAllowKeyboardFocus⚠

igPushButtonRepeat⚠

igPushClipRect⚠

igPushFont⚠

igPushIDInt⚠

igPushIDPtr⚠

igPushIDRange⚠

igPushIDStr⚠

igPushItemWidth⚠

igPushStyleColor⚠

igPushStyleColorU32⚠

igPushStyleVarFloat⚠

igPushStyleVarVec2⚠

igPushTextWrapPos⚠

igRadioButtonBool⚠

igRadioButtonIntPtr⚠

igRender⚠

igResetMouseDragDelta⚠

igSameLine⚠

igSaveIniSettingsToDisk⚠

igSaveIniSettingsToMemory⚠

igSelectable⚠

igSelectableBoolPtr⚠

igSeparator⚠

igSetAllocatorFunctions⚠

igSetClipboardText⚠

igSetColorEditOptions⚠

igSetColumnOffset⚠

igSetColumnWidth⚠

igSetCurrentContext⚠

igSetCursorPos⚠

igSetCursorPosX⚠

igSetCursorPosY⚠

igSetCursorScreenPos⚠

igSetDragDropPayload⚠
Use 'cond' to choose to submit payload on drag start or every frame

igSetItemAllowOverlap⚠

igSetItemDefaultFocus⚠

igSetKeyboardFocusHere⚠

igSetMouseCursor⚠

igSetNextTreeNodeOpen⚠

igSetNextWindowBgAlpha⚠

igSetNextWindowCollapsed⚠

igSetNextWindowContentSize⚠

igSetNextWindowFocus⚠

igSetNextWindowPos⚠

igSetNextWindowSize⚠

igSetNextWindowSizeConstraints⚠

igSetScrollFromPosY⚠

igSetScrollHere⚠	[
Deprecated
]

igSetScrollHereY⚠

igSetScrollX⚠

igSetScrollY⚠

igSetStateStorage⚠

igSetTooltip⚠

igSetWindowCollapsedBool⚠

igSetWindowCollapsedStr⚠

igSetWindowFocus⚠

igSetWindowFocusStr⚠

igSetWindowFontScale⚠

igSetWindowPosStr⚠

igSetWindowPosVec2⚠

igSetWindowSizeStr⚠

igSetWindowSizeVec2⚠

igShowAboutWindow⚠

igShowDemoWindow⚠

igShowFontSelector⚠

igShowMetricsWindow⚠

igShowStyleEditor⚠

igShowStyleSelector⚠

igShowUserGuide⚠

igSliderAngle⚠

igSliderFloat⚠

igSliderFloat2⚠

igSliderFloat3⚠

igSliderFloat4⚠

igSliderInt⚠

igSliderInt2⚠

igSliderInt3⚠

igSliderInt4⚠

igSliderScalar⚠

igSliderScalarN⚠

igSmallButton⚠

igSpacing⚠

igStyleColorsClassic⚠

igStyleColorsDark⚠

igStyleColorsLight⚠

igText⚠

igTextColored⚠

igTextDisabled⚠

igTextUnformatted⚠

igTextWrapped⚠

igTreeAdvanceToLabelPos⚠

igTreeNodeExPtr⚠

igTreeNodeExStr⚠

igTreeNodeExStrStr⚠

igTreeNodePtr⚠

igTreeNodeStr⚠

igTreeNodeStrStr⚠

igTreePop⚠

igTreePushPtr⚠

igTreePushStr⚠

igUnindent⚠

igVSliderFloat⚠

igVSliderInt⚠

igVSliderScalar⚠

igValueBool⚠

igValueFloat⚠

igValueInt⚠

igValueUInt⚠

Type Definitions

ImDrawCallback
Draw callback for advanced use

ImDrawIdx
Vertex index

ImGuiID
Unique ID used by widgets (typically hashed from a stack of string)

ImGuiInputTextCallback
Input text callback for advanced use

ImGuiSizeCallback
Size constraint callback for advanced use

ImTextureID
User data to identify a texture

ImU32
32-bit unsigned integer (typically used to store packed colors)

ImWchar
Character for keyboard input/display

Unions

PairValue
ImGuiStorage value union