Struct imgui_sys::ImGuiStyle

#[repr(C)]

pub struct ImGuiStyle {
    pub alpha: c_float,
    pub window_padding: ImVec2,
    pub window_rounding: c_float,
    pub window_border_size: c_float,
    pub window_min_size: ImVec2,
    pub window_title_align: ImVec2,
    pub child_rounding: c_float,
    pub child_border_size: c_float,
    pub popup_rounding: c_float,
    pub popup_border_size: c_float,
    pub frame_padding: ImVec2,
    pub frame_rounding: c_float,
    pub frame_border_size: c_float,
    pub item_spacing: ImVec2,
    pub item_inner_spacing: ImVec2,
    pub touch_extra_padding: ImVec2,
    pub indent_spacing: c_float,
    pub columns_min_spacing: c_float,
    pub scrollbar_size: c_float,
    pub scrollbar_rounding: c_float,
    pub grab_min_size: c_float,
    pub grab_rounding: c_float,
    pub button_text_align: ImVec2,
    pub display_window_padding: ImVec2,
    pub display_safe_area_padding: ImVec2,
    pub mouse_cursor_scale: c_float,
    pub anti_aliased_lines: bool,
    pub anti_aliased_fill: bool,
    pub curve_tessellation_tol: c_float,
    pub colors: [ImVec4; 43],
}

Runtime data for styling/colors

Fields
alpha: c_float

Global alpha applies to everything in ImGui.

window_padding: ImVec2

Padding within a window.

window_rounding: c_float

Radius of window corners rounding. Set to 0.0 to have rectangular windows.

window_border_size: c_float

Thickness of border around windows. Generally set to 0.0 or 1.0. (Other values are not well tested and more CPU/GPU costly).

window_min_size: ImVec2

Minimum window size. This is a global setting. If you want to constraint individual windows, use igSetNextWindowSizeConstraints().

window_title_align: ImVec2

Alignment for title bar text. Defaults to (0.0, 0.5) for left-aligned, vertically centered.

child_rounding: c_float

Radius of child window corners rounding. Set to 0.0 to have rectangular windows.

child_border_size: c_float

Thickness of border around child windows. Generally set to 0.0 or 1.0. (Other values are not well tested and more CPU/GPU costly).

popup_rounding: c_float

Radius of popup window corners rounding. (Note that tooltip windows use window_rounding)

popup_border_size: c_float

Thickness of border around popup/tooltip windows. Generally set to 0.0 or 1.0. (Other values are not well tested and more CPU/GPU costly).

frame_padding: ImVec2

Padding within a framed rectangle (used by most widgets).

frame_rounding: c_float

Radius of frame corners rounding. Set to 0.0 to have rectangular frame (used by most widgets).

frame_border_size: c_float

Thickness of border around frames. Generally set to 0.0 or 1.0. (Other values are not well tested and more CPU/GPU costly).

item_spacing: ImVec2

Horizontal and vertical spacing between widgets/lines.

item_inner_spacing: ImVec2

Horizontal and vertical spacing between within elements of a composed widget (e.g. a slider and its label).

touch_extra_padding: ImVec2

Expand reactive bounding box for touch-based system where touch position is not accurate enough. Unfortunately we don't sort widgets so priority on overlap will always be given to the first widget. So don't grow this too much!

indent_spacing: c_float

Horizontal indentation when e.g. entering a tree node. Generally == (FontSize + FramePadding.x*2).

columns_min_spacing: c_float

Minimum horizontal spacing between two columns.

scrollbar_size: c_float

Width of the vertical scrollbar, Height of the horizontal scrollbar.

scrollbar_rounding: c_float

Radius of grab corners for scrollbar.

grab_min_size: c_float

Minimum width/height of a grab box for slider/scrollbar.

grab_rounding: c_float

Radius of grabs corners rounding. Set to 0.0 to have rectangular slider grabs.

button_text_align: ImVec2

Alignment of button text when button is larger than text. Defaults to (0.5, 0.5) for horizontally+vertically centered.

display_window_padding: ImVec2

Window position are clamped to be visible within the display area by at least this amount. Only applies to regular windows.

display_safe_area_padding: ImVec2

If you cannot see the edges of your screen (e.g. on a TV) increase the safe area padding. Apply to popups/tooltips as well regular windows. NB: Prefer configuring your TV sets correctly!

mouse_cursor_scale: c_float

Scale software rendered mouse cursor (when io.mouse_draw_cursor is enabled). May be removed later.

anti_aliased_lines: bool

Enable anti-aliasing on lines/borders. Disable if you are really tight on CPU/GPU.

anti_aliased_fill: bool

Enable anti-aliasing on filled shapes (rounded rectangles, circles, etc.)

curve_tessellation_tol: c_float

Tessellation tolerance when using igPathBezierCurveTo() without a specific number of segments. Decrease for highly tessellated curves (higher quality, more polygons), increase to reduce quality.

colors: [ImVec4; 43]

Colors for the user interface